feat: add HomeBelowFold component with HowItWorks, DemoWalkthrough, Features, Testimonials, and FooterCTA sections
feat: implement caching for user resumes and profiles in resumeService

feat: create middleware for session management and redirection based on authentication

chore: enable package import optimization for phosphor-icons in next.config.js
