+++ .github/workflows/ci.yml
@@ -8,6 +8,7 @@ on:
 
 env:
   NODE_VERSION: '24'
+  FORCE_JAVASCRIPT_ACTIONS_TO_NODE24: true
 
 jobs:
   # Linting job - fast feedback on code quality
